core/post-service/pyproject.toml
[tool.poetry]
name = "post-service"
version = "0.1.0"
description = ""
authors = ["wolfychanOwO <102466423+wolfychanOwO@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.71.0"
grpcio-tools = "^1.71.0"
protobuf = "^6.30.2"
requests = "^2.32.3"
pydantic-settings = "^2.9.1"
sqlalchemy = "^2.0.40"
psycopg2 = "^2.9.10"
grpcio-reflection = "^1.71.0"
kafka-python = "^2.1.5"
pytest = "^8.3.5"
alembic = "^1.15.2"
asyncpg = "^0.30.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
